Conflicts:
	drivers/common/openserial.c
	openapps/SConscript
	openstack/02a-MAClow/IEEE802154E.c
	openstack/02b-MAChigh/neighbors.c
	openstack/02b-MAChigh/schedule.h
	openstack/02b-MAChigh/sixtop.c
	openstack/02b-MAChigh/sixtop.h

+++ openstack/02b-MAChigh/neighbors.c
@@ -543,8 +543,9 @@ void neighbors_updateMyDAGrankAndNeighborPreference() {
             rankIncreaseIntermediary = (rankIncreaseIntermediary * 2 * MINHOPRANKINCREASE) / ((uint32_t)neighbors_vars.neighbors[i].numTxACK);
             rankIncrease = (uint16_t)(rankIncreaseIntermediary >> 10);
          }
-
-         tentativeDAGrank = neighbors_vars.neighbors[i].DAGrank+rankIncrease;
+         
+         // cast the uint16_t summands to avoid an overflow if DAGrank == 0xFFFF (MAXDAGRANK)
+         tentativeDAGrank = (uint32_t)neighbors_vars.neighbors[i].DAGrank + (uint32_t)rankIncrease;
          if ( tentativeDAGrank<neighbors_vars.myDAGrank &&
               tentativeDAGrank<MAXDAGRANK) {
             // found better parent, lower my DAGrank
